style: 취향 등록 디자인 변경
- 오모 아이콘 위치 조정
- 챗봇과의 채팅 내용에 대한 간격 조정
